DB refactor of navigational properties, update ReQuesty to use dotnet tool, change a handful of endpoints to work better with DB changes

+++ Frontend/Bones.WebUI/Pages/Asset/ViewAssetPage.razor.cs
@@ -19,7 +19,7 @@ public partial class ViewAssetPage(BonesApiClient apiClient, NavigationManager n
     /// </summary>
     protected bool ApiError { get; set; } = false;
 
-    private int _currentVersion = 0;
+    private long _currentVersion = 0;
     private Guid _projectId = Guid.Empty;
     private string _assetTitle { get; set; } = string.Empty;
